+++ src/main/java/ro/mxp/food/service/RestaurantService.java
@@ -1,5 +1,6 @@
 package ro.mxp.food.service;
 
+import org.jetbrains.annotations.NotNull;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -7,6 +8,7 @@ import org.springframework.stereotype.Service;
 import ro.mxp.food.dto.RestaurantDto;
 import ro.mxp.food.entity.Restaurant;
 import ro.mxp.food.repository.RestaurantRepository;
+import ro.mxp.food.utils.CurrentUsername;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -16,6 +18,9 @@ public class RestaurantService {
 
     @Autowired
     private BCryptPasswordEncoder getBCryptPasswordEncoder;
+
+    @Autowired
+    private CurrentUsername currentUsername;
 
     private ModelMapper modelMapper = new ModelMapper();
 
@@ -32,17 +37,23 @@ public class RestaurantService {
                 .collect(Collectors.toList());
     }
 
-    public void addRestaurant(RestaurantDto restaurantDto) {
+    public void addRestaurant(@NotNull RestaurantDto restaurantDto) {
         restaurantDto.setPassword(getBCryptPasswordEncoder.encode(restaurantDto.getPassword()));
         restaurantRepository.save(modelMapper.map(restaurantDto, Restaurant.class));
     }
 
     public void updateRestaurant(Long id, String email, String username, String restaurantName, String restaurantSpecificity, String restaurantAddress, String restaurantPhone) {
-        restaurantRepository.updateRestaurantRepo(id, email, username, restaurantName, restaurantSpecificity, restaurantAddress, restaurantPhone);
+        Restaurant restaurant = restaurantRepository.findByUsername(currentUsername.displayCurrentUsername());
+        if (restaurant != null && restaurant.getId().equals(id)) {
+            restaurantRepository.updateRestaurantRepo(id, email, username, restaurantName, restaurantSpecificity, restaurantAddress, restaurantPhone);
+        }
     }
 
     public void deleteRestaurant(Long id) {
-        restaurantRepository.deleteById(id);
+        Restaurant restaurant = restaurantRepository.findByUsername(currentUsername.displayCurrentUsername());
+        if (restaurant != null && restaurant.getId().equals(id)) {
+            restaurantRepository.deleteById(id);
+        }
     }
 
 }
